Refactor Selector
