+++ frontend/src/components/index.js
@@ -3,5 +3,6 @@ import Logo from "./Logo";
 import Column from "./Column";
 import Row from "./Row";
 import FlexDiv from "./FlexDiv";
+import Selector from "./Selector";
 
-export { Container, Logo, Column, Row, FlexDiv };
+export { Container, Logo, Column, Row, FlexDiv, Selector };
